Выполнение основной части 4 лабы

+++ tables/employees/index.php
@@ -13,11 +13,22 @@
         <a class="backLinkButton" href="../">
             <img src='/public/icon/back-icon.svg' width='30' height='30' alt='изменить'>
         </a>
-        <a class="button" href="create.php">Создать новую запись</a>
+
+        <?php
+
+        require $_SERVER['DOCUMENT_ROOT'] . '/components/auth.php';
+        authorizationRequired();
+
+        $can_create = rightsCheck('create');
+        if ($can_create) {
+            echo "<a class='button' href='create.php'>Создать новую запись</a>";
+        }
+        ?>
+
     </div>
 
     <?php
-    require '../../components/DBConnect.php';
+    require $_SERVER['DOCUMENT_ROOT'] . '/components/DBConnect.php';
 
     $mysqli = connectToDatabase();
     $sql = "SELECT EMPLOYEES.*, WORKSHOPS.workshop_name 
@@ -31,28 +42,43 @@
                     <th>Имя сотрудника</th>
                     <th>Должность</th>
                     <th>Скидка на одежду</th>
-                    <th>Цех</th>
-                    <th>Изменить</th>
-                    <th>Удалить</th>
-                </tr>";
+                    <th>Цех</th>";
+
+        $can_edit = rightsCheck('edit');
+        $can_delete = rightsCheck('delete');
+
+        if ($can_edit || $can_delete) {
+            echo "                
+                <th>Изменить</th>
+                <th>Удалить</th>";
+        }
+
+        echo "</tr>";
 
         while ($row = mysqli_fetch_assoc($result)) {
             echo "<tr>
                     <td>" . $row['employee_name'] . "</td>
                     <td>" . $row['position'] . "</td>
                     <td>" . $row['discount_on_clothing'] . '%' . "</td>
-                    <td>" . $row['workshop_name'] . "</td>
-                    <td>
+                    <td>" . $row['workshop_name'] . "</td>";
+
+            if ($can_edit) {
+                echo "<td>
                         <a class='iconButton' href='edit.php?id=" . $row['id'] . "'>
                             <img src='/public/icon/edit-icon.svg' width='20' height='20' alt='изменить'>
                         </a>
-                    </td>
-                    <td>
+                      </td>";
+            }
+
+            if ($can_delete) {
+                echo "<td>
                         <a class='iconButton' href='delete.php?id=" . $row['id'] . "' onclick='return confirmDelete();'>
                             <img src='/public/icon/delete-icon.svg' width='20' height='20' alt='удалить'>
                         </a>
-                    </td>
-                </tr>";
+                      </td>";
+            }
+
+            echo "</tr>";
         }
         echo "</table>";
     }
